ecommerceproject/python_backend/pyproject.toml
[tool.poetry]
name = "my-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
